100 - Move ingest_otel_data.py file
